Block User::setPassword from accepting empty password strings
This commit ensures that User::setPassword cannot make a call to the
backend service's setPassword method and provide it an empty string.
While the database backend's already been updated to not accept empty
password strings, this is an extra layer of protection as,
theoretically, other backends may be used, which don't implement this
check. So it avoids the backends from being passed an empty password
string in the first place.