Extract empty password check to a trait
I've done this as I hate unnecessary code duplication, and there seemed
to be no other, reasonable way to achieve it. However, to be honest, I'm
genuinely not sure if this is the right approach to take or not.

I don't want to duplicate the code, so it needs to be created once and
then shared. However, where should it be shared, and how should each
class access it?

It doesn't make sense to create a base class to just contain that one
function, and refactoring both classes to extend that base class would
be a lot of work.

Given that, it makes sense to use a trait and share the function among
the two (and any future) classes. That said, I'm not totally sold on
Traits. Are they a modern implementation of static method calls?

OK, the long and the short of this is that I'm not totally convinced on
Traits yet, but in this case they seem to be the right choice. Should I
change my mind down the track, I want to be able to see why I chose
them.